Renames kicker.ts to dataHelper.ts
