+++ src/fronts/business-today/variantB/components/DefaultCollection.tsx
@@ -6,7 +6,7 @@ import { palette } from "@guardian/src-foundations";
 import { HeadlineCard } from "../../../../components/cards/HeadlineCard";
 import { Multiline } from "../../../../components/Multiline";
 import { Padding } from "../../../../layout/Padding";
-import { kickerText } from "../../../../kicker";
+import { kickerText } from "../../../../dataHelpers";
 import { DefaultCard } from "../../../../components/cards/DefaultCard";
 import { Table, TableRowCell } from "../../../../layout/Table";
