+++ src/fronts/media-briefing/variantB/components/CommentCollection.tsx
@@ -6,7 +6,7 @@ import { Padding } from "../../../../layout/Padding";
 import { Multiline } from "../../../../components/Multiline";
 import { Heading } from "../../../../components/Heading";
 import { HeadlineCard } from "../../../../components/cards/HeadlineCard";
-import { kickerText } from "../../../../kicker";
+import { kickerText } from "../../../../dataHelpers";
 
 export const CommentCollection: React.FC<{
     collection: ICollection;
