+++ src/fronts/media-briefing/variantB/components/TopCollection.tsx
@@ -7,7 +7,7 @@ import { OverlayCard } from "../../../../components/cards/OverlayCard";
 import { Multiline } from "../../../../components/Multiline";
 import { Heading } from "../../../../components/Heading";
 import { HeadlineCard } from "../../../../components/cards/HeadlineCard";
-import { kickerText } from "../../../../kicker";
+import { kickerText } from "../../../../dataHelpers";
 
 export const TopCollection: React.FC<{
     collection: ICollection;
